tried aligning with huskylens in auto, disabled some opmodes
