+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Tests/MotorTest.java
@@ -29,6 +29,7 @@
 
 package org.firstinspires.ftc.teamcode.Tests;
 
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -50,6 +51,7 @@ import com.qualcomm.robotcore.util.ElapsedTime;
  */
 
 @TeleOp(name = "Motor Test", group = "test")
+@Disabled
 
 public class MotorTest extends LinearOpMode {
 //    private Gyroscope imu;
