+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Tests/RoadrunnerTeleOpTest.java
@@ -30,6 +30,7 @@
 package org.firstinspires.ftc.teamcode.Tests;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
@@ -40,6 +41,7 @@ import org.firstinspires.ftc.teamcode.drive.DriveConstants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 
 @TeleOp(name = "RoadrunnerTeleOpTest", group = "Test")
+@Disabled
 
 public class RoadrunnerTeleOpTest extends LinearOpMode {
     private DcMotor backLeft;
